refactor(Direction): 선택 지점까지의 방향 계산 코드 Direction 클래스로 이동
