+++ src/main/java/domain/piece/King.java
@@ -5,7 +5,7 @@ import domain.direction.Direction;
 import domain.position.Position;
 import java.util.List;
 
-public class King extends SpecificLocationPiece {
+public class King extends Piece {
 
     private static final List<Direction> DIRECTIONS = List.of(
         Direction.NORTHEAST,
@@ -40,9 +40,9 @@ public class King extends SpecificLocationPiece {
     @Override
     protected Direction direction(Position source, Position target) {
         Direction direction = super.direction(source, target);
-        if(!DIRECTIONS.contains(direction)){
+        if (!DIRECTIONS.contains(direction)) {
             throw new IllegalArgumentException();
         }
-        return  direction;
+        return direction;
     }
 }
