+++ src/main/java/domain/piece/Queen.java
@@ -4,7 +4,7 @@ import domain.Player;
 import domain.direction.Direction;
 import java.util.List;
 
-public class Queen extends MovableRangePiece {
+public class Queen extends Piece {
 
     private static final List<Direction> DIRECTIONS = List.of(
         Direction.NORTHEAST,
